pyproject.toml
[project]
name = "ds-star-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=1.2.10",
    "langgraph>=0.6.11",
    "langchain-openai>=1.1.8",
    "langchain-anthropic>=1.1.0",
    "langchain-google-genai>=2.1.0",
    "langchain-ollama>=1.0.0",
    "python-dotenv>=1.2.1",
]
